feature- 출고 상태 문제 해결 05_23 17:48

+++ src/main/java/co/second/easyrp/chart/web/ChartController.java
@@ -54,7 +54,7 @@ public class ChartController {
     	List<ChartVO> DonutData = chartService.getTop3Product(donut_year, month);
     	
     	List<String> labels = new ArrayList<>();
-    	List<Integer> data = new ArrayList<>();
+    	List<Long> data = new ArrayList<>();
     	
     	for (ChartVO chartVO : DonutData) {
     		labels.add(chartVO.getProdname());
